scan-backups: process each volume at separate Thread, in parallel
Each time a disk is inserted, spawn a new thread to scan its contents,
thus be able to read multiple drives in parallel.

Note that "eject()" must now be part of that Thread, so that it is
executed at its end (rather than wait()ing for the thread or some
other elaborate algo).